% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dir}
\alias{dir}
\title{Inspect global namespace or package namespace}
\usage{
dir(package)
}
\arguments{
\item{package}{package name to inspect}
}
\value{
vector of function and method names in the namespace
}
\description{
If called without an argument, return the names in the current scope.
If called with a package name, return the names in the package scope.
}

