% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not in}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{the values to be matched.}

\item{y}{the values to be matched against.}
}
\value{
logical vector
}
\description{
Opposite of `%in%`
}

