% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%|\%}
\alias{\%|\%}
\title{Maybe do monad}
\usage{
f \%|\% g
}
\arguments{
\item{f}{left hand function}

\item{g}{right hand function}
}
\description{
Use operator to evaluate functions from left to right, breaking and
returning NA if any function fails or returns NA.
}

