% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{last}
\alias{last}
\title{Last element}
\usage{
last(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments to be passed to tail}
}
\value{
last element in x
}
\description{
Take the last element in `x`
}

