% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{len}
\alias{len}
\title{Length}
\usage{
len(x)
}
\arguments{
\item{x}{vector}
}
\value{
integer indicating the number of characters in x if x is a
single-element character vector, otherwise the number of elements in x
}
\description{
Length of `x` if `x` is a vector, number of characters if `x` is a string
}

