% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{not}
\alias{not}
\title{Check `x` is a valid value}
\usage{
not(x)
}
\arguments{
\item{x}{value}
}
\value{
logical
}
\description{
Similar to the `not` keyword in python. not(NA) will evaluate to FALSE,
whereas not(NULL) evaluates to TRUE.
}

