% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{q}
\alias{q}
\title{Quit}
\usage{
q(save = "no", ...)
}
\arguments{
\item{save}{save the current session? yes/no}

\item{...}{passed to `quit`}
}
\value{
exits R
}
\description{
Quit without save prompt
}

