% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rename}
\alias{rename}
\alias{rename<-}
\title{Rename object names}
\usage{
rename(x, old, newn)

rename(x) <- value
}
\arguments{
\item{x}{object to rename}

\item{old}{names to match on}

\item{newn}{new names, matched to `old` by position}

\item{value}{list with `old` and `new` containing the names to match on and
names to be matched to, by position}
}
\description{
Match and rename object names with new names.
}
\examples{
x <- data.frame(foo = 1:3, bar = letters[1:3])
names(x) <- rename(x, "bar", "baz")
names(x)
# [1] "foo" "baz"
rename(x) <- list(old = "baz", new = "bar")
names(x)
# [1] "foo" "bar"
}

