% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_na_rows}
\alias{rm_na_rows}
\title{Remove rows with NA}
\usage{
rm_na_rows(x)
}
\arguments{
\item{x}{data.frame to be cleaned}
}
\value{
data.frame with zero or more rows removed
}
\description{
Remove all rows that have on or more NA values
}

