% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{section}
\alias{section}
\title{Split a vector by index}
\usage{
section(x, n = 50L)
}
\arguments{
\item{x}{vector to split.}

\item{n}{section length.}
}
\value{
list of vector sections
}
\description{
Split a vector into sections of equal length
}

