% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{unfactor}
\alias{unfactor}
\title{Change factors to characters}
\usage{
unfactor(x)
}
\arguments{
\item{x}{object to change factors into characters in}
}
\description{
Change factors to characters
}
\examples{
df <- data.frame(a = letters[1:5], x = 1:5, y = LETTERS[1:5], stringsAsFactors = TRUE)
str(df)
df <- unfactor(df)
str(df)
}

