% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{last}
\alias{last}
\title{Last `n` elements in `x`}
\usage{
last(x, n = 1L)
}
\arguments{
\item{x}{list or vector}

\item{n}{number of elements to take from the end}
}
\value{
last n elements in x
}
\description{
Last `n` elements in `x`
}

